requirements.txt
streamlit
openai
python-dotenv
xhtml2pdf
markdown